+++ src/main/java/com/hobbyhop/domain/user/service/KakaoService.java
@@ -1,179 +1,7 @@
 package com.hobbyhop.domain.user.service;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.hobbyhop.domain.user.enums.UserRoleEnum;
-import com.hobbyhop.domain.user.dto.KakaoUserInfoDTO;
-import com.hobbyhop.domain.user.entity.User;
-import com.hobbyhop.domain.user.repository.UserRepository;
-import com.hobbyhop.global.exception.user.JsonProcessException;
-import com.hobbyhop.global.security.jwt.JwtUtil;
 import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.RequestEntity;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
-import org.springframework.web.client.RestTemplate;
-import org.springframework.web.util.UriComponentsBuilder;
 
-import java.net.URI;
-import java.util.UUID;
-
-@Slf4j(topic = "KAKAO Login")
-@Service
-@RequiredArgsConstructor
-public class KakaoService {
-
-    private final PasswordEncoder passwordEncoder;
-    private final UserRepository userRepository;
-    private final RestTemplate restTemplate;
-    private final JwtUtil jwtUtil;
-
-    public void kakaoLogin(String code, HttpServletResponse response){
-        // 1. "인가 코드"로 "액세스 토큰" 요청
-        String accessToken = getToken(code);
-
-        // 2. 토큰으로 카카오 API 호출 : "액세스 토큰"으로 "카카오 사용자 정보" 가져오기
-        KakaoUserInfoDTO kakaoUserInfo = getKakaoUserInfo(accessToken);
-
-        // 3. 필요시에 회원가입
-        User kakaoUser = registerKakaoUserIfNeeded(kakaoUserInfo);
-
-        // 4. JWT 토큰 반환
-
-        // accessToken 생성
-        String username = kakaoUser.getUsername();
-        String userAccessToken = jwtUtil.createAccessToken(username);
-        // accessToken 을 클라이언트에게 헤더로 넣어 보냄
-        response.setHeader("Authorization", userAccessToken);
-        // username을 key로, accessToken을 value로 -> redis에 저장
-        jwtUtil.saveAccessTokenByUsername(username, userAccessToken);
-
-        // refreshToken 생성
-        String refreshToken = jwtUtil.createRefreshToken(username);
-        // accessToken을 key로, refreshToken을 value로 -> redis에 저장
-        jwtUtil.saveRefreshTokenByAccessToken(userAccessToken, refreshToken);
-    }
-
-    private String getToken(String code){
-        // 요청 URL 만들기
-        URI uri = UriComponentsBuilder
-                .fromUriString("https://kauth.kakao.com")
-                .path("/oauth/token")
-                .encode()
-                .build()
-                .toUri();
-
-        // HTTP Header 생성
-        HttpHeaders headers = new HttpHeaders();
-        headers.add("Content-type", "application/x-www-form-urlencoded;charset=utf-8");
-
-        // HTTP Body 생성
-        MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
-        body.add("grant_type", "authorization_code");
-        body.add("client_id", "6666e4d3e7955e33eac0eb2e6609e3e5"); // REST API 키 6666e4d3e7955e33eac0eb2e6609e3e5
-        body.add("redirect_uri", "http://localhost:8080/api/users/login/kakao/callback");
-        body.add("code", code);
-
-        RequestEntity<MultiValueMap<String, String>> requestEntity = RequestEntity
-                .post(uri)
-                .headers(headers)
-                .body(body);
-
-        // HTTP 요청 보내기
-        ResponseEntity<String> response = restTemplate.exchange(
-                requestEntity,
-                String.class
-        );
-
-        // HTTP 응답 (JSON) -> 액세스 토큰 파싱
-        JsonNode jsonNode;
-        try {
-            jsonNode = new ObjectMapper().readTree(response.getBody());
-        } catch (Exception e) {
-            throw new JsonProcessException();
-        }
-        return jsonNode.get("access_token").asText();
-    }
-
-    private KakaoUserInfoDTO getKakaoUserInfo(String accessToken) {
-        // 요청 URL 만들기
-        URI uri = UriComponentsBuilder
-                .fromUriString("https://kapi.kakao.com")
-                .path("/v2/user/me")
-                .encode()
-                .build()
-                .toUri();
-
-        // HTTP Header 생성
-        HttpHeaders headers = new HttpHeaders();
-        headers.add("Authorization", "Bearer " + accessToken);
-        headers.add("Content-type", "application/x-www-form-urlencoded;charset=utf-8");
-
-        RequestEntity<MultiValueMap<String, String>> requestEntity = RequestEntity
-                .post(uri)
-                .headers(headers)
-                .body(new LinkedMultiValueMap<>());
-
-        // HTTP 요청 보내기
-        ResponseEntity<String> response = restTemplate.exchange(
-                requestEntity,
-                String.class
-        );
-
-        JsonNode jsonNode;
-        try {
-            jsonNode = new ObjectMapper().readTree(response.getBody());
-        } catch (Exception e) {
-            throw new JsonProcessException();
-        }
-        Long id = jsonNode.get("id").asLong();
-        String nickname = jsonNode.get("properties")
-                .get("nickname").asText();
-        String email = jsonNode.get("kakao_account")
-                .get("email").asText();
-
-        log.info("카카오 사용자 정보: " + id + ", " + nickname + ", " + email);
-        return new KakaoUserInfoDTO(id, nickname, email);
-    }
-
-    private User registerKakaoUserIfNeeded(KakaoUserInfoDTO kakaoUserInfo) {
-        // DB 에 중복된 Kakao Id 가 있는지 확인
-        Long kakaoId = kakaoUserInfo.getId();
-        User kakaoUser = userRepository.findByKakaoId(kakaoId).orElse(null);
-
-        if (kakaoUser == null) {
-            // 카카오 사용자 email 동일한 email 가진 회원이 있는지 확인
-            String kakaoEmail = kakaoUserInfo.getEmail();
-            User sameEmailUser = userRepository.findByEmail(kakaoEmail).orElse(null);
-            if (sameEmailUser != null) {
-                kakaoUser = sameEmailUser;
-                // 기존 회원정보에 카카오 Id 추가
-                kakaoUser.kakaoIdUpdate(kakaoId);
-            } else {
-                // 신규 회원가입
-                // password: random UUID
-                String password = UUID.randomUUID().toString();
-                String encodedPassword = passwordEncoder.encode(password);
-
-                // email: kakao email
-                String email = kakaoUserInfo.getEmail();
-
-                kakaoUser = User.builder()
-                        .username(kakaoUserInfo.getNickname())
-                        .password(encodedPassword)
-                        .email(email)
-                        .role(UserRoleEnum.USER)
-                        .kakaoId(kakaoId)
-                        .build();
-            }
-            userRepository.save(kakaoUser);
-        }
-        return kakaoUser;
-    }
+public interface KakaoService {
+    public void kakaoLogin(String code, HttpServletResponse response);
 }
